Country change docCode from int to string

+++ src/main/java/ru/bellintegrator/practice/dao/impl/UserDAOImpl.java
@@ -119,7 +119,7 @@ public class UserDAOImpl implements UserDAO {
             objConjunction.add((Restrictions.eq("doc.code", userFilterView.getDocCode().intValue())));
         }
         if (userFilterView.getCitizenShipCode() != null) {
-            objConjunction.add(Restrictions.eq("country.code", userFilterView.citizenShipCode.intValue()));
+            objConjunction.add(Restrictions.eq("country.code", userFilterView.citizenShipCode));
         }
 
         cr.add(objConjunction);
